報告ボタンを修正 https://github.com/aipictors/aipictors/issues/88
