+++ app/[lang]/(main)/works/[work]/_components/share-popover.tsx
@@ -9,6 +9,7 @@ import {
 import { usePathname } from "next/navigation"
 import { RiFileCopyLine, RiShareFill } from "react-icons/ri"
 import { toast } from "sonner"
+import CopyWorkUrlButton from "./copy-work-url"
 import { XIntent } from "./share-on-x"
 
 type Props = {
@@ -17,16 +18,6 @@ type Props = {
 
 export const SharePopover = (props: Props) => {
   const currentUrl = `https://www.aipictors.com${usePathname()}`
-
-  const handleCopyUrl = async () => {
-    try {
-      await navigator.clipboard.writeText(currentUrl)
-      toast("URLがクリップボードにコピーされました。")
-    } catch (err) {
-      toast("URLのコピーに失敗しました。")
-      console.error(err)
-    }
-  }
 
   return (
     <Popover>
@@ -41,14 +32,7 @@ export const SharePopover = (props: Props) => {
             <h4 className="font-medium leading-none">イラストを共有する</h4>
           </div>
           <div className="grid gap-2">
-            <Button
-              className="flex items-center gap-2"
-              variant="outline"
-              onClick={handleCopyUrl}
-            >
-              <RiFileCopyLine />
-              URLをコピー
-            </Button>
+            <CopyWorkUrlButton currentUrl={currentUrl} />
             <XIntent
               text={`${props.title}\n`}
               url={`${currentUrl}\n`}
@@ -57,7 +41,9 @@ export const SharePopover = (props: Props) => {
             <Button
               className="flex items-center gap-2"
               variant="outline"
-              onClick={handleCopyUrl}
+              onClick={() => {
+                toast("開発中につき、未実装")
+              }}
             >
               <RiFileCopyLine />
               イラストをコピー
